services/social-dispatcher/requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-cloud-tasks
pydantic
